Conflicts:
	include/bolt/amp/detail/scan.inl
	include/bolt/amp/scan.h
	include/bolt/btbb/detail/scan.inl
	include/bolt/btbb/detail/scan_by_key.inl
	include/bolt/cl/bolt.h
	include/bolt/cl/detail/copy.inl
	include/bolt/cl/detail/count.inl
	include/bolt/cl/detail/fill.inl
	include/bolt/cl/detail/inner_product.inl
	include/bolt/cl/detail/reduce.inl
	include/bolt/cl/detail/reduce_by_key.inl
	include/bolt/cl/detail/scan.inl
	include/bolt/cl/detail/scan_by_key.inl
	include/bolt/cl/detail/sort.inl
	include/bolt/cl/detail/sort_by_key.inl
	include/bolt/cl/detail/stablesort_by_key.inl
	include/bolt/cl/detail/transform.inl
	include/bolt/cl/detail/transform_reduce.inl
	include/bolt/cl/detail/transform_scan.inl
	include/bolt/cl/device_vector.h
	include/bolt/cl/functional.h
	include/bolt/cl/scan_by_key_kernels.cl
	test/amp/ScanTest/ScanTest.cpp
	test/cl/CopyTest/CopyTest.cpp
	test/cl/DeviceVectorTest/device.vector.test.cpp
	test/cl/FillTest/FillTest.cpp
	test/cl/GenerateTest/GenerateTest.cpp
	test/cl/InnerProductTest/InnerProductTest.cpp
	test/cl/ReduceByKeyTest/ReduceByKeyTest.cpp
	test/cl/ReduceTest/ReduceTest.cpp
	test/cl/ScanByKeyTest/ScanByKeyTest.cpp
	test/cl/ScanTest/ScanTest.cpp
	test/cl/SortByKeyTest/SortByKeyTest.cpp
	test/cl/SortTest/SortTest.cpp
	test/cl/StableSortByKeyTest/StableSortByKeyTest.cpp
	test/cl/StableSortTest/StableSortTest.cpp
	test/cl/TransformReduceTest/TransformReduceTest.cpp
	test/cl/TransformScanTest/TransformScanTest.cpp
	test/cl/TransformTest/TransformTest.cpp

+++ include/bolt/AsyncProfiler.h
@@ -27,7 +27,7 @@ class AsyncProfiler
 {
 private:
     LARGE_INTEGER constructionTimeStamp;
-    size_t timerPeriodNs;
+    LONGLONG timerPeriodNs;
     size_t numThrowAwayTrials;
 
 public:
